Bug fix #3029298: Default Id generation limited to 127 elements

+++ commons/trunk/src/main/java/org/databene/commons/NumberUtil.java
@@ -52,7 +52,7 @@ public class NumberUtil {
 				int.class,        Short.MAX_VALUE, 
 				Integer.class,    Integer.MAX_VALUE, 
 				long.class,       Integer.MAX_VALUE, 
-				Long.class,       Byte.MAX_VALUE, 
+				Long.class,       Long.MAX_VALUE, 
 				float.class,      Float.MAX_VALUE, 
 				Float.class,      Float.MAX_VALUE, 
 				double.class,     Double.MAX_VALUE,
